requirements.txt
langgraph 
langchain 
langchain-google-genai
python-dotenv
ipython